Decimal scores for measurements and slap halving
- ScoreEntry.Points and Slap.Penalty are now double; scoreboard + event standings totals are double
  (question points stay integer); DecimalScores migration. Uses double (SQLite REAL) so SUM/ORDER work
- Score games can record decimals (e.g. 12.5 s, 3.5 m); BouleBoard measurement input accepts decimals
- Slap halving is now exact (e.g. gap 1 -> 0.5, no flooring); penalty/transfer carry the fraction
- Shared Fmt.Num trims whole numbers and shows halves; applied to scoreboard, standings and podium
- Verified: 12.5 s measurement on the scoreboard; odd-gap slap -> .5 penalty (187 -> 186.5)

+++ src/Rundan.Server/Data/Entities/ScoreEntry.cs
@@ -19,7 +19,9 @@ public class ScoreEntry
     public User? User { get; set; }
 
     public int Round { get; set; } = 1;
-    public int Points { get; set; }
+
+    /// <summary>The recorded value (points, seconds, or millimetres). Decimal for measured games.</summary>
+    public double Points { get; set; }
     public string? Note { get; set; }
     public DateTimeOffset RecordedUtc { get; set; }
 }
